refactor(home page features): rename folders to change case (step 1 of 2)
In order to change folder case, I need to first rename to something else then rename again back to
original with correct case.
