+++ src/features/HomePage/components/hero-section2/index.js
